refactor(sdk): split monolithic firecrawl.go into modular file structure
- Split 838-line firecrawl.go into 9 domain-specific files
- client.go: struct, constructor, headers
- types.go: all request/response type definitions
- scrape.go, crawl.go, map.go, search.go: endpoint methods
- errors.go, helpers.go, options.go: internal utilities
- Zero logic changes — pure structural refactor
